+++ GameServer/Action.java
@@ -1,0 +1,6 @@
+package GameServer;
+
+public abstract class Action {
+    abstract Response executeAction(Request request);
+
+}
