+++ client/jungle-app/src/component/Square.js
@@ -16,7 +16,7 @@ class Square extends React.Component {
                 value={this.props.row + "," + this.props.column}
                 className={"game-buttons " + (this.props.available ? "availableSpace" : "notAvailableSpace")}
                 onClick={() => {
-                    this.props.postExample("action=move_piece&gameID=1234&username=" + this.props.username + "&password=iforgot123&row=" + this.props.row + "&column=" + this.props.column)
+                    this.props.postExample("action=MovePiece&gameID=1234&username=" + this.props.username + "&password=iforgot123&row=" + this.props.row + "&column=" + this.props.column)
                 }}
                 // onClick={this.props.postExample()}
             >{this.props.piece}.{(this.props.environment != "empty" ? this.props.environment : "")}.{this.props.available}</button>
